Move EventType LogField to the analytics module.
