+++ src/app_engine/apprtc.py
@@ -361,7 +361,7 @@ def add_client_to_room(request, room_id, client_id, is_loopback):
           %(client_id, room_id, retries))
 
       if room.get_occupancy() == 2:
-        analytics.report_event(constants.EventType.ROOM_SIZE_2,
+        analytics.report_event(analytics.EventType.ROOM_SIZE_2,
                                room_id,
                                host=request.host)
       success = True
